[webhook] integrate webhook as a module
