+++ webhook/cmd/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"os"
+
+	webhook "github.com/starlinglab/integrity-v2/webhook"
+)
+
+func main() {
+	webhook.Run(os.Args[1:])
+}
