+++ webhook/file.go
@@ -1,4 +1,4 @@
-package utils
+package webhook
 
 import (
 	"io"
@@ -8,7 +8,7 @@ import (
 
 var outputDirectory = os.Getenv("FILE_OUTPUT_PATH")
 
-func CopyOutputToFile(src io.Reader, originalFileName string, cid string) error {
+func CopyOutputToFilePath(src io.Reader, originalFileName string, cid string) error {
 	ext := filepath.Ext(originalFileName)
 	path := filepath.Join(outputDirectory, cid)
 	if ext != "" {
